Memory management
Ensure we delete the container in the tests to kick off the child
shared_ptr chain. Once you hand an object to the display list, the
display list controls the life cycle of that object. (unless you
explicitly regain control of your object via the returned shared
pointer to addChild)