SharedPointer Interface
The display list explicitly requires that we have the ability to create
a shared_ptr from ourselves. In order to facilitate this feature I am
going to run with this experimental flair::make_shared interface that
will ensure that we have a weak_ptr to ourselves constructed from the
first shared_ptr that is handed out. I feel like this use is better
than the std::enable_shared_from_this since I do not want to have to do
any multiple inheritance magic on a per class basis.